refactor: 옵션과 카테고리의 상품 조회 기능 상품 컨트롤러로 위치 변경
